+++ skeletons/der_encoder.c
@@ -118,7 +118,7 @@ der_write_TL(ber_tlv_tag_t tag, ber_tlv_len_t len,
 
 	/* Serialize tag (T from TLV) into possibly zero-length buffer */
 	tmp = der_tlv_tag_serialize(tag, buf, buf_size);
-	if(tmp == -1 || tmp > sizeof(buf)) return -1;
+	if(tmp == -1 || tmp > (ssize_t)sizeof(buf)) return -1;
 	size += tmp;
 
 	/* Serialize length (L from TLV) into possibly zero-length buffer */
